Move export coordinates of track to library
